refactor(docs): spinner dx
