+++ apps/docs/content/components/spinner/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Spinner} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Spinner color="default"/>
-      <Spinner color="primary"/>
-      <Spinner color="secondary"/>
-      <Spinner color="success"/>
-      <Spinner color="warning"/>
-      <Spinner color="danger"/>
-    </div> 
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
