+++ apps/docs/content/components/spinner/label-colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Spinner} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Spinner label="Default" color="default" labelColor="foreground"/>
-      <Spinner label="Primary" color="primary" labelColor="primary"/>
-      <Spinner label="Secondary" color="secondary" labelColor="secondary"/>
-      <Spinner label="Success" color="success" labelColor="success"/>
-      <Spinner label="Warning" color="warning" labelColor="warning"/>
-      <Spinner label="Danger" color="danger" labelColor="danger"/>
-    </div> 
-  );
-}`;
+import App from "./label-colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
