+++ apps/docs/content/components/spinner/sizes.raw.jsx
@@ -1,0 +1,11 @@
+import {Spinner} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4">
+      <Spinner size="sm" />
+      <Spinner size="md" />
+      <Spinner size="lg" />
+    </div>
+  );
+}
