+++ apps/docs/content/components/spinner/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Spinner} from "@nextui-org/react";
+
+export default function App() {
+  return <Spinner />;
+}
